server/requirements.txt
fastapi
fastmcp
uvicorn
torch
matplotlib
pandas
polars
chronos-forecasting
anyio
trio
ollmcp
